🚚 Move ISCN edit functionality to modal component
